[#1825] use int in ifindex_'s getters and setters
to match the type of ifindex_

+++ src/lib/dhcp/iface_mgr.h
@@ -213,7 +213,7 @@ public:
     /// @brief Returns interface index.
     ///
     /// @return interface index
-    int64_t getIndex() const { return ifindex_; }
+    int getIndex() const { return ifindex_; }
 
     /// @brief Returns interface name.
     ///
@@ -400,7 +400,7 @@ protected:
     std::string name_;
 
     /// Interface index (a value that uniquely identifies an interface).
-    int64_t ifindex_;
+    int ifindex_;
 
     /// List of assigned addresses.
     AddressCollection addrs_;
@@ -487,7 +487,7 @@ public:
             boost::multi_index::hashed_unique<
                 // Use the interface index as the key.
                 boost::multi_index::const_mem_fun<
-                    Iface, int64_t, &Iface::getIndex
+                    Iface, int, &Iface::getIndex
                 >
             >,
             // Start definition of index #2.
